+++ packages/dara-components/dara/components/graphs/components/causal_graph_viewer.py
@@ -24,6 +24,7 @@ from dara.components.graphs.components.base_graph_component import BaseGraphComp
 from dara.components.graphs.definitions import EditorMode
 from dara.components.graphs.graph_layout import PlanarLayout
 from dara.core.interactivity import DerivedVariable, Variable
+from dara.core.logging import dev_logger
 
 
 class CausalGraphViewer(BaseGraphComponent):
@@ -226,7 +227,7 @@ class CausalGraphViewer(BaseGraphComponent):
     js_module = '@darajs/components'
 
     causal_graph: Optional[Union[CausalGraph, DerivedVariable, Variable]]
-    editor_mode: EditorMode = EditorMode.DEFAULT
+    editor_mode: Optional[EditorMode] = None
 
     @validator('causal_graph')
     @classmethod
@@ -250,6 +251,8 @@ class CausalGraphViewer(BaseGraphComponent):
             isinstance(values.get('graph_layout'), PlanarLayout)
             and values.get('editor_mode', EditorMode.DEFAULT) != EditorMode.DEFAULT
         ):
-            raise ValueError('Planar Layout is currently only supported with EditorMode.DEFAULT.')
-
+            dev_logger.warning(
+                'Planar Layout is currently only supported with EditorMode.DEFAULT. Setting editor_mode to EditorMode.DEFAULT.'
+            )
+            values['editor_mode'] = EditorMode.DEFAULT
         return values